added return type for valid moves of peices
