+++ app/src/main/java/src/Chessboard.java
@@ -1,16 +1,12 @@
 package src;
 
-import java.time.Period;
-import java.util.ArrayList;
-import java.util.List;
+import src.Pieces.Piece;
 
 public class Chessboard {
     public final static int DIMENSION = 8;
-    private Piece[][] board;
+    private static Piece[][] board = new Piece[DIMENSION][DIMENSION];;
     private Team whiteTeam;
     private Team blackTeam;
-    private Player player1;
-    private Player player2;
     boolean turn;
     public Chessboard() {
         initializeChessboard();
@@ -21,10 +17,6 @@ public class Chessboard {
         blackTeam = new BlackTeam();
         blackTeam.initializeTeam();
     }
-    public void gameInit() {
-        printBoard();
-    }
-
     /**
      * Prints the board in console
      */
@@ -39,7 +31,7 @@ public class Chessboard {
                     piece = blackTeam.getPiece(col,row);
                 }
                 if (piece != null) {
-                    String figure = String.valueOf(piece.getFigure()) + " ";
+                    String figure = String.valueOf(piece.getColor()) + String.valueOf(piece.getFigure()) + " ";
                     System.out.print(figure);
                 }
             }
