+++ app/src/main/java/src/Main.java
@@ -3,6 +3,5 @@ package src;
 public class Main {
     public static void main(String[] args) {
             Chessboard chessboard = new Chessboard();
-            chessboard.gameInit();
     }
 }
