+++ app/src/main/java/src/Team.java
@@ -1,5 +1,7 @@
 package src;
 
+import src.Pieces.Piece;
+
 import java.util.List;
 import java.util.ArrayList;
